Construct filters using floats instead of ints

+++ csComp/directives/Widgets/Filters/BarFilterCtrl.ts
@@ -102,7 +102,7 @@ module Filters {
                 if (!d.properties.hasOwnProperty(filter.property)) return null;
                 else {
                     if (d.properties[filter.property] != null) {
-                        var a = parseInt(d.properties[filter.property]);
+                        var a = parseFloat(d.properties[filter.property]);
                         if (a >= min && a <= max) {
                             return Math.floor(a / binWidth) * binWidth;
                         } else {
@@ -162,7 +162,7 @@ module Filters {
 
             filterFrom.on('change', () => {
                 if ($.isNumeric(filterFrom.val())) {
-                    var min = parseInt(filterFrom.val());
+                    var min = parseFloat(filterFrom.val());
                     var filters = this.dcChart.filters();
                     if (filters.length > 0) {
                         filters[0][0] = min;
@@ -176,7 +176,7 @@ module Filters {
             });
             filterTo.on('change', () => {
                 if ($.isNumeric(filterTo.val())) {
-                    var max = parseInt(filterTo.val());
+                    var max = parseFloat(filterTo.val());
                     var filters = this.dcChart.filters();
                     if (filters.length > 0) {
                         filters[0][1] = max;
